Beacon board update - add buffer for standalone and peripheral modes
Added a buffer between the stack and the GPS / LoRa radio, controlled by an onboard MSP (replacing the 555 timer). This causes the LoRa and GPS to either be controlled by the stack or by the onboard MSP depending on the operating mode. The stack pinout changed slightly to match the v2.1 stack pinout.